+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 Changes:
 
+- Ensure `Range/RangeInclusive` variant check does not fail on encoding
 - Ensure non-option calls in api-contract are marked as `@deprecated`